separate 2 files from nextauth-route.ts
